+++ Cargo.toml
@@ -13,7 +13,6 @@ path = "src/main.rs"
 [dependencies]
 serde_json = "^1.0.57"
 uuid = { version = "0.8.2", features = ["serde", "v1"] }
-blake2b_simd = "0.5.11"
 pbr = "1"
 tonic = "0.5.2"
 indradb-lib = { path = "indradb/lib" }
@@ -24,10 +23,9 @@ bzip2 = "0.4.3"
 quick-xml = "0.22.0"
 regex = "1"
 serde = { version = "1", features = ["derive"] }
-bincode = "1.3.3"
 lazy_static = "1"
 tokio = { version = "1.11.0", features = ["rt-multi-thread"] }
 warp = "0.3.2"
-failure = "0.1.8"
 async-channel = "1.6.1"
 tera = "1"
+chrono = { version = "0.4.19" }
