+++ Makefile
@@ -11,12 +11,7 @@ data/enwiki-latest-pages-articles.xml.bz2:
 	mkdir -p data
 	cd data && wget 'https://dumps.wikimedia.org/enwiki/latest/enwiki-latest-pages-articles.xml.bz2'
 
-data/archive_dump.bincode: data/enwiki-latest-pages-articles.xml.bz2
-	time cargo run --release -- parse \
-		--archive-path data/enwiki-latest-pages-articles.xml.bz2 \
-		--dump-path data/archive_dump.bincode
-
-data/wikipedia.rdb: data/archive_dump.bincode indradb/target/release/indradb-server
+data/wikipedia.rdb: data/enwiki-latest-pages-articles.xml.bz2 indradb/target/release/indradb-server
 	time cargo run --release -- index \
 		--dump-path data/archive_dump.bincode \
 		--database-path data/wikipedia.rdb
